Cleans and Fixes
@jkaberg - this fixes the calls to CP and SB for the lack of an API call
option in SB.  Also fixes some parameter information for SB, and
generally cleans up the verbiage.

However...when both CP and SB are set to True to be active, it is ALWAYS
sending the CP information to process_media (starts line 268) instead of
actually identifying both parameters in that if/and statement are true.